requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
babel==2.18.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
build==1.4.0
cachetools==7.0.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.0
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
courlan==1.3.2
cryptography==46.0.5
csvw==3.7.0
ctransformers==0.2.27
curl_cffi==0.14.0
dateparser==1.3.0
ddgs==9.10.0
distro==1.9.0
dlinfo==2.0.0
duckduckgo_search==8.1.1
durationpy==0.10
edge-tts==7.2.7
espeakng-loader==0.2.4
fake-useragent==2.2.0
fastapi==0.99.1
filelock==3.21.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
kokoro-onnx==0.4.7
kubernetes==35.0.0
language-tags==1.2.0
lxml==6.0.2
lxml_html_clean==0.4.4
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mem0ai==0.0.1
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
narwhals==2.17.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu12==11.4.1.4
nvidia-nvjitlink-cu12==12.9.86
oauthlib==3.3.1
ollama==0.3.3
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==24.2
pandas==2.3.3
phonemizer-fork==3.3.1
pillow==12.1.1
pinecone==8.1.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
pip_system_certs==5.3
posthog==5.4.0
primp==1.1.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
protoc-gen-openapiv2==0.0.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
segments==2.3.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soundfile==0.13.1
soupsieve==2.8.3
sse-starlette==3.0.3
starlette==0.27.0
streamlit==1.55.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
textblob==0.19.0
tld==0.13.2
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
trafilatura==2.0.0
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
yarl==1.22.0
zipp==3.23.0
